Data Engineering Challenges
This challenge is used to extract data/web-scrap from two different tools; selenium and beautiful soup.
for being an avionics engineer to achieve this task and get the 1000 best results is not easy. I am not a software engineer/IT engineer and didn't have any frequent connections in this field but still accepted the challenge and took initiative with Scratch and start-up by reading even a tiny piece of information regarding the programming side because I am a beginner in this programming world.
I was not able to achieve the goal but try my level best to do that task, just because I am still in progress to do that task and now I am finding myself pretty self-satisfied that at least I done and accepted that task and completed all the parameters of that task now I'm enjoying more to learn this further and able to make myself more sufficient in this programming world. There were a lot of errors I faced during coding even I wrote the code 6,7 times to make it better and mistakeless.